requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
nibabel==5.1.0
pydicom==2.4.3
numpy==1.26.2
pillow==10.1.0
python-jose==3.3.0
requests 
fastapi
jinja2
nibabel
numpy
pillow
pydicom
python-multipart
uvicorn